pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paku"
version = "0.1.0"
description = "Instagram screenshot data pipeline"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "Pillow>=10.0",
    "PyYAML>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
]

[project.scripts]
paku = "paku.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["paku"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
